# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Hostel Leave & Digital Outpass Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── LEAVE REQUESTS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, tutor, hod, principal, warden, security, super_admin]
        registerNumber:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        hostelRoom:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        hostelBlock:
          type: ["string", "null"]
        parentName:
          type: ["string", "null"]
        parentWhatsapp:
          type: ["string", "null"]
        parentEmail:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, tutor, hod, principal, warden, security, super_admin]
        registerNumber:
          type: string
        departmentId:
          type: integer
        hostelRoom:
          type: string
        phone:
          type: string
        parentPhone:
          type: string
        photoUrl:
          type: string
        classId:
          type: integer
        hostelBlock:
          type: string
        parentName:
          type: string
        parentWhatsapp:
          type: string
        parentEmail:
          type: string
        address:
          type: string
        designation:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        parentPhone:
          type: string
        hostelRoom:
          type: string
        departmentId:
          type: integer
        photoUrl:
          type: string
        classId:
          type: integer
        hostelBlock:
          type: string
        parentName:
          type: string
        parentWhatsapp:
          type: string
        parentEmail:
          type: string
        address:
          type: string
        designation:
          type: string

    # ── Leave Request ─────────────────────────────────────────────────────────
...